#15 [Rename] 프로젝트 디렉토리 구조 리팩토링
