+++ TEST/rendering_test.py
